`fs_update` can now take a list of files #35
If one article id is given and the created article is already of type
'fileset', but more than one file is given, they will all be uploaded
to that article id.

If more than one article id is given, and a list of files is given,
then each id will get the matching file in the list.

See the documentation ?fs_upload for details.

Method needs testing and test cases written.  should close #35